// lib/api/generated/models/loadSpec.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { PctOf } from './pctOf';

export type LoadSpec = {
  kg: number;
  mode: 'kg';
} | {
  mode: 'pct';
  of: PctOf;
  pct: number;
} | {
  mode: 'bar';
} | {
  mode: 'athlete';
};