// lib/api/generated/models/pctOf.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type PctOf = typeof PctOf[keyof typeof PctOf];


export const PctOf = {
  snatch: 'snatch',
  clean_jerk: 'clean_jerk',
  total: 'total',
  exercise: 'exercise',
} as const;